New option to not depend on OpenSSL (Experimental)
This option is enabled with the USE_CURL_PUBLIC_KEY_PINNING compilation
flag.
It's needed to put the files with public keys of SSL certs that you can
find in the folder contrib/publickeys in the working folder of the app.
According with cURL documentation (version 7.41.0) this option should
work for cURL builds of version 7.39.0 with OpenSSL, GnuTLS or GSKit.