Fixes for the management of inshares without a sharekey
With these changes, the SDK will receive those inshares as NO_KEY
Both cases has been fixed, when the inshare without a key is received in
an action packet, and when it's received in a full reload.
When the key is restored, it's applied and all decrypted nodes are saved
to the database.