Amended ReadMe.md for Sphinx set up
Jaimes had attempted to create the sphinx documentation and found he was unable to as the MAKEFILE for sphinx was missing from the folder. This is due to the file not being uploaded to github as there is no file ext for it and so needs to be generated by the sphinx set up before the documentation can be created.