Async queries - Initial commit
New objects:
- DbQuery: represents a database query to be executed by DbThread
- DbQueryQueue: thread-safe queue to handle DbQueries
- DbThread: includes the loop function to execute queries

Multi-thread logic:
The `DbThread` is started by `MegaClient`. When the SDK receives a
request of type `TYPE_GET_NUM_CHILD_FOLDERS/FILES` from the `MegaApi`,
the SDK creates and adds the new query to the queue and notifies
(`DbThread::notify()`) the `DbThread` to wake it up and start the
execution.
The results from `DbThread` are sent to the `MegaApi` directly from the
`DbThread` by calling the corresponding callback (such as
`MegaClient::getnumchildfolders_result(int, error)`).

Others:
- New `MegaApp` callbacks: `void getnumchildfiles_result(int, error)`
and `void getnumchildfolders_result(int, error)`
- Definitions of `MegaMutex` and `MegaThread` have been moved from
`MegaApi` to SDK (they're now used for `DbThread`)
- Added the interface `Waiter::notify()` to the abstract class