Manage pending references
`NodesCache` is now able to manage references to nodes that are still in
use, so they should not be deleted from the cache if it's full.

The new algorithm tries to find and erase references that are not in use
until the cache's size is decreased below the limit. If it is not
possible, then it waits for a number of new additions (that is the
situation when references are erased) and tries again.