Fix public links deletion
Deletion of the outshare of a folder automatically deletes any public
link in the outshare, so the request to delete the link was not sent.
It's not convenient since the client apps don't receive the
`onRequestFinish()` as expected.
Now, when deleting a link, first the command to delete the link is sent,
and then the command to delete the outshare.