Removed unneeded checks in syncdown() - sync_syncable() is already called before adding nodes to nchildren - Nodes on nchildren always have a 'n' attribute
Set node->localnode to ~0 to point out that the localnode is being overwritten
If it's not going to be overwritten but there is something in the target path,
the download is aborted until the path is properly scanned

Do not try to fix persistent errors creating folders. If the folder already exists,
it should be scanned shortly and that would fix the problem.
On any other use case (for example, a problem with permissions),
previous hacks wouldn't help anyway