Revamping of user attribute put/get methods
- Removal of the specificities of private attributes (prefixed by `*`):
every attribute is considered private, even those without that prefix,
like `country`. Others are whitelisted, so contacts are able to get the
value, like `firstname`.

In summary, it's better to use the full name of attributes internally,
instead of trying to add a different prefix depending on the type of
attribute, since the prefix doesn't always match with the scope.

- Addition of new user attributes: `*!lstint` and `*!authring`: prepare
the SDK to get/put those attributes gracefully.

- Fix the removal of user's avatar image: the value is "none" (without
base64 encoding), so it requires specific treatment to work.