Runtime caching of user names
Since the `firstname` and `lastname` are requested very often from SDK
apps, they are now stored as part of the `User` information when
retrieved by first time. As a consequence, the information is available
without any API request.
If the information has changed, it's simply invalidated (so next time it
will require the API request to get the updated value.