Improve initialization of keyring
The attribute should not be retrieved at
`CommandFetchNodes::procresult()`, but after the reception of the wait
URL, since the full fetchnodes process doesn't finish with the command,
but after processing the pending actionpackets (due to fetchnodes
caching feature in API).
Also, at session resume, if the keyring is not cached, launch the
initialization.