Fix sharing with registered non-contact emails
A share can be set for a contact, but also for non-contacts. In that
case, the email provided may or may not be registered in MEGA.

When the email is registered, but not a contact yet, a previous public
RSA key request returns the userhandle (which is stored in the User
object) and, therefore, the `uid` becomes the handle instead of the
email (regardless the user is not a contact yet).

All the above has a side effect on the `s2` request. The API assumes
that requests with handles are for contacts and requests with emails
require to create the contact request. So for registered emails but non-
contacts yet, the email should be sent.