Fix ephemereal user
`megacli` creates an ephemereal account as the first step to create a
full account.

Now, the API returns `"m":"none"` for the email of the ephereal user. It
makes the SDK to store the email and identify the account as confirmed,
instead of ephemereal, blocking the signup process from the SDK (in
`megacli`).