Support for new user attribute
The `!^prd` attribute is used to store useful information for a
_Password Reminder Dialog_, which aims to reduce the number of users who
eventually forget their password and, as consequence, lose access to
their accounts.
This attribute stores information like whether the Recovery Key was
exported ever or not, or when the user logged in by typing his/her
password by last time.