Fix result tag for putua()
When the `prd` user attribute is set internally, it requires 2 chained
requests, first one to get the value, second one to set the updated
value. The resulting request tag is wrongly set for the second request,
resulting in a request that never finishes. However, the request is kept
in the request's map until a logout, when it's forced to finish with
EACCESS.
Now, with this fix, the request is correctly finished and notified to
the app via `onRequestFinish()`