Ignore ua packets based on ua version
API will eventually start sending in user-attribute actionpackets the
most recent version of the attribute, rather than every single update.
That change will also remove the `i` in the APs, so clients will need to
filter ua packets based on the version. If the version is known, then
the ua was set by the client and can ignore the AP.