use fs::u8path instead of construct from string
This fixes some cases for utf8 where the fs::path constructor was not preserving the utf8 text.
Also adjusted the CMakeLists for consistent use of mega_glob.h/cpp.  If we use one we should use the other too.
Also moved the cmake specific config.h.in.cmake file to a cmake-only folder.