Recent Actions feature.
This feature is being added to the SDK to match the feature in the webclient, available from the clock on the left hand bar.
You will be able to view changes to your MEGA account, as groups of files that were recently uploaded or recently updated.
A different user, folder or significant time difference would mean the files go into a different bucket.