Add method to catch up with API
In order to ensure the client is up to date with the server in terms of
reception of recent changes (actionpackets), this method closes existing
sc connection (if any) and abort the backoff to immediately retrieve
changes.
Since the app may require to check this situation several times, app's
requests are queued and sent sequentially.