cloudraid: use partial http responses, fix progress callbacks
If an http request for file data fails, but already pulled back a megabyte of data, it's a shame to waste that.
Now we detect that case, and process it as if it was a successful request for the amount of data that actually arrived.
Also the progress callbacks had an incorrect position indicator for raid files, that is now fixed.