Enable putting commands in separate batches for server execution.
Reworked the Request queue which used to have two batches and send them alternately.  Commands can be marked to be put in their own batch, or for the bactch to be sent without SID (for the misc flags command).
Some of the Request queue logic is formatted with an eye to processing the response items gradually, for a future feature (that this code is adapted from).