RequestDispatcher: fix `logout` response calling `clear()` on active structure
Since RequestDispatcher::clear() is called from one of the items managed by the dispatcher, and clear() wipes the entire data structure,
delay that wiping until the function returns (and avoid calling any further command responses in the meantime).
Once we can be sure that none of that data structure is in use anywhere, we perform the clear().
(also fixed a merge issue with the CMakeLists.txt)