updates from review
Simplify specifying encryption of the whole file (use -1 for length), length is now int64_t.
Check mediainfo data in analyseMediaInfo (and allow files to be uploaded without it, eg. photos)
Update for autocomplete to be able to specify quoted "-1" on megacli/megaclc command line, and have it interpreted as numeric parameter rather than flag.