Fix decryption of attrs in `pli` response
We need to use the key provided in the public folder link to be able to
decrypt the attributes, so we need to provide it to
getPublicLinkInformation. On the other hand with the full link we can
check if the link is a valid folder link, otherwise the request will
fail.

Fix heap-use-after-free bug by removing an unecesary call to
fireOnRequestFinish