Don't overwrite masterkey with folderkey
Do not use `folderaccess()` for retrieval of public link info, since it
sets the key of the folder as masterkey, and also the rootnode to be the
public handle of the folder and the authentication token.

Do not include the `folderlinkinfo_result()` inside the `ENABLE_CHAT`
macro, since it's not specific to chat.

Improve error's management: missing/invalid key, decryption error...