improved algorithm for preventing recursion
We want to avoid that uploading a path (originlocalpath) writes
something into a MEGA location synced with a descendant of that path.
With one exception: the upload entails the update of a corresponding
local path with its synced MEGA node.
  Common use case 1: upload excluded files
  Common use case 2: we a have a local structure and the same in MEGA,
and we have something synchronize, but some other folders arent. We want
to upload to have a manual one-way up-sync.