Patch TLV with invalid record
When Webclient resets the authentication ring for a contact and there is
no more contacts in the account, it write a TLV record with `T=\0`,
which is illegal. For compatibility, the TLV library will return the TLV
with the records read until that point of the container, instead of a
null pointer.