Adjust xxtea Encrypt/Decrypt for reactions
Depending on how we acquire the xxtea input (i.e byte-by-byte) we need
to adjust the Byte order format like in media file attributes. In case
of reactions the input of xxtea is acquired directly from memory so we
don't need to adjust the Byte order format and we have to skip this
step.