Enable SDK clients to lock the sdkMutex for performance reasons
Sometimes clients need to make many calls in a row that each lock the sdkMutex.  If the SDK is busy, many interruptions may occur in that sequence.   This allows that to occur all in one quick go.

MEGAsync in particular can benefit form this because it does a sequence of locking in its paint functions when transfers are first processed there.

The object returned manages the lock on the mutex, ensuring it is locked max once (at a time), and is unlocked on object destruction.