Add auto transaction management to SQLite wrapper class
Due to database edits occurring in transfer and file destructors, we can't use parameters to ensure that programmers have instantiated a DBTableTransactionCommitter first.
Instead, now the database class will check that with an assert to remind us that one is needed (if a new flag is set on the db table instance when it is created).
Additionally, the db table class will auto begin the transaction on the first edit, so no need to call beginOnce() manually anymore.
This version still keeps the DBTableTransactionCommitter parameters where we can, to remind us to set those up (transaction will be committed in its destructor).