Transfer and TransferSlot backoff timer efficency.
Only check the backoff timers that we need to when preparing a wait.   We used to iterate the whole list of transferslots and transfers checking each one, which cost a lot of CPU when the system was busy.
This change maintains a set of just those timers that are in play, and only checks that set (for each of the cases:  slots, put transfers, get transfers).
Also address a bug with the `retrying` flag not getting set true when a slot needs to back off, and a bug where that backoff was being set to a very large value when it shouldn't (which didn't cause any problem due to the first bug)