performance improvement: don't create and destroy a SymmCipher on every call, which can be very frequent
In a test downloading 2500 tiny files, the SymmCipher was not actually used, but took 18% CPU.
Moved the variable to the object where it can be reused, rather than creating on the stack every time.
Also updated the cmake project so the generated visual studio project doesn't have to be manually updated to enable profiling.