Prevent applykeys() taking CPU when all node keys are resolved.
If all nodes have their keys, then there is no need to scan through the entire node map, which we were doing quite frequently, which slowed things down a lot when processing actionpackets etc.
Separated `nodekey`out from NodeCore, so that we can track updates to it separately in Node.  NewNode which used to inherit that field, gets its own field with that name instead.
Node's version of `nodeky`is now private, called `nodekeydata`, with a read accessor.  Functions that update it also update the count of applied keys.
If the number of nodes that have keys applied matches the number of nodes (taking into account that the root nodes do not have keys), then the applykey() loop is skipped.
Also added a megacli command to exercise `setmaxconnections` (related to transfer performance testing).