Avoid unnecessary fopen/fingerprinting, fix MegaApi_impl shutdown threading issue
fixed threading issue: Previously the MegaClient was destroyed on the Impl thread, but then the Impl destructor might need it when deleting any remaining request objects.
Fixed unnecessary fingerprint: MegaFilePut needs one and megasync will do it if it's not done already.  And the code there already has fingerprinted the file once, we can just use it.
Fixed unnecessary fopen(): onFolderAvailable iterates the folder for uploads, but opened each entry to find out if it was a file or folder.  The entry can supply that info though.
refactor: Made removeCaches private, and called from locallogout as that one was only used before locallogout, and an upcoming change will add code that must be done before that, and for all locallogout()s.