Improve tctable update times when executing many same-type requests
Start one DBTableTransactionCommitter that will be used for consecutive loops in sendPendingRequests().
The transaction is committed if we exit the loop, process 1024 requests, or a different type of request is processed (and a new transaction will be started if there is any tctable activity)
This improvement makes a really huge difference for things like pausing or cancelling all transfers when there are a large number of them.  Those actions are almost instant now (in combination with some other PRs under review).