Fix  unserialize of non-syncable localnodes
If a record happens to not _hasSyncable_, the `ptr` is not moved after
the byte that it actually writes at serialization --> `assert(ptr ==
end)` will fail.
Furthermore, the byte for the `hasSyncable` is actually not needed,
since only new records will contain the extra byte for the flag and the
extension bytes. Old records will not enter into the block `if (ptr <
end)`, since actually `ptr = end`.