Allow fastLogin to resume offline
eg. for music player apps that need to be able to browse the node tree, but already have downloaded songs to play.

Offline means that the SC subsystem is not active, and server requests are rejected with EARGS.

Should the app need to go online again later, it can use fastLogin again without setting the offline flag.