use unique_ptr to guarantee ownership of FA data
Make sure that the compiler would complain if the string containing the file attribute data was changed between the API call and the metad call.
Also make sure that the string is deleted correctly.
There was no functional problem, but knowing the compiler was checking those things would have made a recent investigation easier.