Send file versioning flag request, and unshareable key request before fetchnodes
Because fetchnodes is now batched by itself, we want to make sure that those fields are available by the time fetchnodes completes, when the app might use them.
Sending them before fetchnodes is sent will solve that, when it fetches data from the servers.
When data is fetched from cache though, that happens in the login_result callback, and data is loaded immediately, with fetchnodes_result called directly.
That bypasses any requests still in the request queue.   The effect is consistent with starting up in the old state, and then receiving updates for everything that changed.