Avoid to invalidate attributes of new contacts
When an unknonw user becomes a new contact, there are no attributes to
be invalidated. Hence, apps should not rely on the `changed` bitmask to
update user attributes (which might be cached by the app, even when a
user is not a contact. In example, a non-contact in a groupchat)