Store LocalNode shortnames in the localnode database.
The SDK was looking up the shortnames on the loading of a Sync, which doesn't take much time for a small sync but for a million node sync takes around 4 minutes (on SSD).
This PR stores those in the database so they can be retrieved much more quickly on startup - the million node sync localnodes load in about 10 seconds.
The first time a sync is loaded after this change, the localnodes still have to have the shortname looked up, so the first time is still slow.  The localnodes are written back to the database periodically (after every 50000).
Also the shortname is not stored in RAM anymore if it is the same as the localnode's localname, saving some memory.