Don't overwrite error code of logout
An automatic logout may happen in result of 1) a HTTP request error, or
2) when the account gets blocked. In both cases, the logout is notified
as failed with the corresponding error (the http error or the
`API_EBLOCKED`).
In order to notify the apps properly, the logout should return `API_OK`,
but indicate the reason of the logout in the
`MegaRequest::getParamType()`.