Address warnings from logging.h when trying to build test_integration for Mac
Some types were mismatched and caused lots of signed/unsigned warnings (repeating since they were in the header)
Also found lots of areas where the ENABLE_LOG_PERFORMANCE was not checked, and various projects failed to build if that was turned on.
Removed Logging_test.cpp code since that was not building with that on, and we are using and checking logging all the time anyway.
Added some casts in net.cpp to ensure we use the size_t version of DirectMessage() constructor, and highlight that we are truncating the size argument at that point (for those targets that have 32 bit size_t).
Use std::chrono::seconds(4) syntax instead of 4s which is c++14 (for Mac)