Added "stronger" specialization of nodebyfingerprint(...).
This new specialization performs a more detailed check when determining
whether it's safe to create a remote copy in the cloud instead of
uploading some local file.

The function determines whether it is safe to create such a copy by
computing a metamac of the local file using the remote node's encryption
key. If the metamac matches, the files are considered identical.

If the metamac cannot be generated, the files are considered distinct.
This seems to the safest approximation as it trades bandwidth for file
integrity.