Fix for cs requests that take more than 2 minutes to transmit
The server lock check mechanism was activating, and determining that the servers were not busy, so the request must have finished/failed without us receiving a reply.
Setting the lastdata field on the request as we are transmitting the request prevents this issue.
Added a couple of warning fixes in that file also.