Fix for crypto on threads in the Transfer destructor
When a Transfer is destroyed, we should process and write any data already downloaded so as to not waste the user's bandwidth.
Also those downloaded chunks must wait for decryption to finish if it is in progress, and decrypt any chunks downloaded but not yet decrypted.

Also the AsyncQueue is adjusted to have a flag as to whether operations can be discarded on MegaClient destruction (which is one case for transfer deletion)
Decryption of downloads are not discardable, but encrypt for uploads are.
If the MegaClient is destroyed, we let the decryptions run, but discard the encryptions.

The threading issue is addressed because we no longer try to merge the mac map of a still-decrypting chunk.