Adding cryptopp-staticcrt dependency
We need a version of cryptopp built statically, and with static CRT, for MEGAUpdater.  Having it built in the same vcpkg triplet makes MEGAsync configuration much easier.
So cryptopp-staticcrt is just a copy of the cryptopp port, and a few small changes will be made to build it against the static CRT.