When a file is renamed, its shortname may not have been changed
In this case, if we are scanning the tree because file notifications got overloaded, then checking the old LocalNode's presence by testing shortname gives the wrong answer.
That resulted in re-downloading the file when only uploads should have been happening, and possibly having the file duplicated locally and remotely with both names.
Address this with two parts
  - larger buffers for the file notifications so we get all the notifications for large updates (or at least make failure less common)
  - make that presence check with the full names, not with short names.  Microsoft is deprecating short names anyway, so let's use the most common case whenever we can.