Added a chain of pending commands to Node
Node attributes are not updated before sending the change request to the servers
SetAttr command uses those to make sure multiple commands contain all locally pending changes
Commands can generate their JSON just before sending (so SetAttr can account for failed commands, retries, etc)
Fixed logout sequence in case the request is made while other commands are pending
syncdown took a bool parameter but it was only ever true, removed
Deal with tricky backwards compatible issue: move command produces d and t packets, and t cannot have the st field, or we break old clients.
In syncdown, if we have started sending a change to the servers for a node, do not re-process it yet
In syncup, if we already connected a localnode with a node, don't make a new node - this will need to be expanded on
Added basic timing to see why SdkTest suite is so slow - mostly setup/teardown.