Merged LocalPath, and used it to fix sync cases for drives such as F:\
Squash-merged the feature/local-path-class branch, keeping enough of it to add LocalPath to the code base, and use it for fixing this sync case.
On windows, syncing of drive paths such as F:\ which have to end in \, had trouble because the path addition ended up with paths like F:\\folder which are illegal.
Now LocalPath takes care of concatenating those paths, and getting those cases right.
Additionally, it simplifies the code considerably by factoring out fiddly string manipulation, and worries about which platform you're on.
Additionally, it abstracts away the windows case where we are storing two-byte characters in std::string, so now we have a type tracking whether the string contains those.
Previously it was very easy to mis-use a string, passing a two-byte character string to a function expecting UTF-8 or vice versa.
But the main problem fixed is that now you can set up a sync from F:\ to eg. /F in the cloud.