Testing for *scsn could send uninitialized data, fixed now
The main fix is for scsn to be fully 0'd on MegaClient construction.
Also the setting of *scsn=0 and stopsc=true were pretty much the same condition.
Wrapped up the scsn and its validity in a small helper class which can also validate its state when used.
In future perhaps btsc could also be added to that class as it is also highly interdependent with it.
One small compiler warning fixed in net.cpp