For the MassNotifyFromLocalFolderTree test, wait for notifications to finish processing
On a slower machine, it may take longer, the main thing is that we do process them all eventually, without losing any
So we check the queue lengths periodically, waiting until they are all finished, with an overall timeout of 5 minutes.
Also fixed a shutdown issue where destroying a Sync was slow unless we were also destroying the client, which is the
case for this test, so adjusted that optimization to apply to ~Sync.