Fix upload nonresponseiveness due to checking filesystem typeComparator
For a large folder upload, minimize checking of which filesystem we are working in.  Just once per folder.
Took advantage of the new LocalPath to simplify whether the path we are checking is utf8 or utf16 (for windows)
getlocalfstype and getFilesystemType work on LocalPath now, which means fewer switches between utf8/utf16
that means we can use LocalPath functions and don't need getValidPath anymore
Upload functions in megaApi_impl take a FS parameter, for cases where we already know.  If it is unknown then the FS is looked up.