Combined refactoring changes from SIC removal branch
Refactoring in a separate PR before submitting SIC removal proper.
Commands are passed Result which already checked for errors, and lets the command know if an array or object has been entered.
Try a completion function option for unlink()
Putnodes takes a vector<NewNode> instead of pointers and counters.
In putnodes_result, the caller can swap() to get back ownership of the NewNodes if needed.