Add constant for max extension length
getextension methods have a parameter to specify max extension length.
In current code, this parameter is set to 8 characters, in most of the
places where method is called, and some sopported extensions like
`ecelp4800` are longer than 8 chars.

We are going to add a new constant to unify this max extension len, and
make easier to modify it, if longer extensions are added in the future.