Completion functions instead of share_result() and exportnode_result()
Inspired by the need to fix the failing test case for getpubliclink().
This enables us to easily code the test cases for export/share/getpubliclink()
And also makes the chaining of commands very simple - export node needs to share and then getpubliclink().
But those are not always called in that order.   Matching up the response to the request
and then figuring out whether to chain another command was complex.
Now we can use lambdas to supply the actions to perform when a command completes, with no need to match up tags and request types.
This change can serve as an example for discussion as to whether we should adopt the method widely.
Also without needing to use overrides of MegaApp virtual functions, we can keep the requesting and response processing code close together.