Process all subtransfers in one shot
If we are processing a folder to be uploaded/downloaded, we want that
all subtransfers in transfersQueue are processed and cached in one
shot, to avoid that part of them are not cached if user closes the app.

We also need to set a max timeout after wich we will unlock mutex,
otherwise if folder tree that we want to process is big enough, app
could be blocked for a long time.