Implement mechanism to avoid infinite loops upon failed putnodes
If the target of putnodes is removed from cloud drive, upon reception
of command by API, it will fail with API_EAGAIN (-3), and command will
be retried with an exponencial backoff in an infinite loop.

To avoid this issue, we need to check if current request has failed
with -3, if it's of type TYPE_CREATE_FOLDER_TREE and target doesn't
exists anymore. In this case we need to make request fail.