Map temp nodeHandle to definitive one
If we want to create a folder tree with a single putnodes, we need to
specify the entire folder structure, so we need to provide temp
nodeHandles to match each node with it's parent.

When we read received nodes from API, we need to match temp nodeHandle
to definitive one, to be able to find the parentNode of pending files
to be uploaded.