Provide optional completion function to CommandPutNodes
When we create a folder tree in one shot, maybe we need to send
multiple Putnodes, one per different target. To avoid process a
request, and in order to simplify the logic, we can add new
CommandPutNodes directly, and provide a completion function that will
process the results of the commands.

In case that completion function is not provided, putnodes_result will
be called by default, to preserve the original behaviour.