SDK-1293. Only use shortname for FS notification matching
Any calls to the filesystem, use the long name.  It'll be the same code path for systems with and without (which is getting much more common) having shortname generation turned on.
Shortname will only be used for filesystem notification matching.
Also fixed some warnings (on windows) that had crept into the code base.