Add new flag to MegaTransferPrivate
The `startChatForUpload()` used the flag for streaming, which should be
in use only for downloads. However, there's generic code that takes into
account that flag for downloads too. In result, transfers started with
the aforementioned method cannot be cancelled and are returned by
`MegaApi::getStreamingTransfers()`.

This commits adds a new flag to the transfers, so it can be used for
this type of uploads without interference with streaming transfers.