Enqueue the update of backup names to avoid race condition
When multiple backups are registered/removed at the same time, their
backup names will be sent all together. Since the corresponding user's
attribute is versioned, the API will only accept the first `upv`,
replying `API_EEXPIRED` for the rest of the `upv` of subsequent backups.

The solution in this commit is to enqueue any update if there's already
an update in flight. Also, it optimizes the number of required `upv`
commands, since all the pending backup names will be sent together,
instead of individually.