No changes in the backup names are expected
With the introduction of the Backup Center (which may remove backups and
their names), as well as with a lazy initialization of the backup name
upon registration of the backup (the app may fail to set the name), it
may happen that updates of the corresponding attribute actually have
nothing to update. This commit aims to cover those cases, avoiding the
error log message and the assertion.