Fix for assertion failure during non-logged-in get
Also fixed win32 cwd() so it doesn't have a trailing \0, so that database path construction works and we do create a tctable, like on linux.
Adjusted committers so that nesting is ok as in this case - both in checkfile_result and Request::process.
Also tidied up related raw pointers with unique_ptr.