Enable session resume for writable folder links.
Session resume is also enabled for read-only folder links.
Resuming means we can load the locally cached node database instead of re-requesting from the server, just like a normal account
Of course for folder links it's not really using a session id, but follows the pattern of calling dumpsession() and re-login with the output of that.