fix issues with remote paths changes in syncs
updates on remote paths for syncs are only considered for active syncs
after sync revamping. This reverts that and makes notice of inactive
syncs changes in remote paths. Thus allows for apps to notice that and
also be able to re-enable syncs whose remote went to the //bin an back