Fix conflicts between complete and cancel
We need to notify app about parent transfer in cancel method, otherwise
in some cases, as if we cancel operation before worker thread starts
running we will lose that notification.

On the other hand we have to take into account that cancel method is
called from MegaTransferPrivate destructor, which will be invoked from
complete method, so we need to guarantee that MegaTransferPrivate
destructor is called only once.

For this reason we need an atomic variable that prevents to execute
cancel or complete if the other action has previously executed. With
this fix we will prevent double free issues.

Ensure that there are no subtransfers if we are calling complete
method.