Add ensureSyncUserAttributes()
This function is to make sure we always get the right user attributes that contain
the keys etc associated with storing sync configs.  There is only one version per user
of that attribute, and we need to make sure we cope with any race to create it.
Each client must end up with the one true value.  Once it's received, it can be
subsequently retrieved from cache, in the User objects.