Add completion function to `ug` call
If we pass a nullptr completion function here, the method may start the
`putua` (if `mOnEnsureSyncUserAttributesComplete` is still null). Then,
the next call to `ensureSyncUserAttributes(completion)` made by the
request/s `TYPE_COPY_SYNC_CONFIG` will start a new `putua`, which we
don't want to.
Of course, the second `putua` will fail with `API_EEXPIRED`, and will
trigger the `getua` (unneeded, since the UA is already available
locally).
This commit aims to avoid that roundtrip by passing a simple completion
function.