Avoid to add syncs to vector in copySyncConfig()
This method is used to populate the SyncConfigs cache with the ones
historically stored/managed by the app (MEGAsync). The copied
SyncConfigs will be restored from cache upon
`resumeResumableSyncsOnStartup()`, adding them to the
`Syncs::mSyncsVec`. Adding them to the vector before that point results
on duplicated items in the vector (apart from the dupe failing to start,
since both share the same path).