Always queue the original path and LocalNode for RETRY
For notifications on Windows and Mac, the LocalNode is always the sync
root node, and it's only Linux that varies from that - for Linux, the
LocalNode pointer is always to the immediate parent of the item being
notified.
By keeping the parameters the same for a RETRY, we can be sure that
remains the case.