SDK-1440. Extend the freeimage patch to fix a discovered compiler bug
Also removed the megacmd and megasync specific preferred-ports files as those are in their respective repos now.

The compiler bug is in platformXBM inside freeimage.   It sets up a lookup table, writing all 256 values to it first, and then updating some entries.  With optimization on in the msvc compiler version 19, (eg. 19.28.29913 from Visual Studio 2019 16.9.2), the first value set after the loop, at index 0x30, is not set to 0 as the code attempts, but left at 256.    This causes the XMB files to be mis-interpreted, and falls into an endless loop (due to reaching an undetected EOF - the patch includes EOF detection too).