Untangle exportnode() and setshare().
The logic for these is quite simple but the implementation with callbacks made it quite confusing.  exportnode() was actually setting shares, not exporting the node, and relying on the called back intermediate layer (or direct app) to pick up from there and finish the job, and actually export the node.   Anyway this has all now been straightened out, using completion functions to clearly express what the next step should be for the various cases (setshare is not always needed).

These are changes from the Sync Rework branch (they help clarify the logic for these commands with the SIC removal changes), which cause merge issues when merging develop to that branch, it will be good to update develop with these changes and avoid those merge issues in future.

exportnode() (for folders) first calls setshare() and then if that completes successfully, requests a public link for it.
megacli used to have to have separate logic for that in its share_result override, as did the intermediate layer
now the SDK core takes care of it in a straightforward way, without all the fiddling about with reqtag/restag, figuring out what the request type is in callbacks, and coping with overloadded parameters stored in the same MegaRequest fields for different request types.